requirements.txt
Flask==3.0.0
pytest==7.4.0
openai==1.54.3
mock==5.1.0
python-dotenv==1.0.1